back/requirements.txt
Django==5.2
requests
asgiref
sqlparse
djangorestframework
gunicorn
# psycopg2
drf-spectacular
python-dotenv